Move Webhook GET method to AddForm POST method.
